flink-autoscaler: added implementation to query KafkaMessagesPerSeconds from Prometheus and publishs it to Kafka in the metric topic for testing flink-autoscaler: chore: added mapstruct library for mapper implementation to avoid boilerplate code

+++ flink-autoscaler/src/main/java/com/mpds/flinkautoscaler/domain/model/events/MetricReported.java
@@ -20,7 +20,7 @@ public class MetricReported extends DomainEvent {
 
     private final String flinkTopic;
 
-    private final int kafkaMessagesPerSeconds;
+    private final float kafkaMessagesPerSeconds;
 
     private final int kafkaMaxMessageLatency;
 
@@ -37,7 +37,7 @@ public class MetricReported extends DomainEvent {
     private final float memoryUsage;
 
     @JsonCreator
-    public MetricReported(@JsonProperty("sequenceNumber") Long sequenceNumber, @JsonProperty("kafkaMessagesInPerSeconds") int kafkaMessagesPerSeconds, @JsonProperty("occurredOn") LocalDateTime occurredOn, @JsonProperty("kafkaTopic") String kafkaTopic, @JsonProperty("flinkTopic") String flinkTopic,
+    public MetricReported(@JsonProperty("sequenceNumber") Long sequenceNumber, @JsonProperty("kafkaMessagesInPerSeconds") float kafkaMessagesPerSeconds, @JsonProperty("occurredOn") LocalDateTime occurredOn, @JsonProperty("kafkaTopic") String kafkaTopic, @JsonProperty("flinkTopic") String flinkTopic,
                           @JsonProperty("kafkaMaxMessageLatency") int kafkaMaxMessageLatency, @JsonProperty("recordsProcessedPerSeconds") int recordsProcessedPerSeconds,  @JsonProperty("networkInPerSeconds") float networkInPerSeconds, @JsonProperty("networkOutPerSeconds") float networkOutPerSeconds,
                           @JsonProperty("sinkHealthy") Boolean sinkHealthy, @JsonProperty("cpuUtilization") float cpuUtilization, @JsonProperty("memoryUsage") float memoryUsage) {
         super(UUID.randomUUID(), sequenceNumber, occurredOn);
